+++ packages/base/core/CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+* Switched from native EventEmitter to custom class that is based on EventTarget
+
 ## [1.0.0] - 2024-12-16
 
 ### Added